refactor: AppHeader에 하드코딩된 메뉴 데이터 리팩토링
