+++ 003 Code/FE/notion-linked-blog/src/components/common/MenuItem.tsx
@@ -1,0 +1,16 @@
+import {MenuItemHrefs} from "@/types";
+import {Typography} from "antd";
+import Link from "next/link";
+
+const {Text} = Typography;
+
+interface ItemProps {
+	index: number;
+	value: string;
+}
+
+export default function MenuItem({index, value}: ItemProps) {
+	return (
+		<Link href={MenuItemHrefs[index]}><Text>{value}</Text></Link>
+	);
+}
